chore: renamed component Resizable to ResizablePanel
