+++ view/desktop/src/components/ResizablePanel.tsx
